Update Dao_NhanVien.java
luong

+++ src/dao/Dao_NhanVien.java
@@ -1,5 +1,6 @@
 package dao;
 
+import java.math.BigDecimal;
 import java.sql.CallableStatement;
 import java.sql.Connection;
 import java.sql.Date;
@@ -245,7 +246,7 @@ public class Dao_NhanVien {
 				String email = rs.getString(7);
 				boolean gioiTinh = rs.getBoolean(8);
 				boolean chucVu = rs.getBoolean(9);
-				double luong = rs.getDouble(10);
+				BigDecimal luong = rs.getBigDecimal(10);
 				NhanVien nv = new NhanVien(maNV, hoTen, mk, ngaySinh, diaChi, soDienThoai, email, chucVu, gioiTinh,
 						luong);
 				dsnv.add(nv);
@@ -271,7 +272,7 @@ public class Dao_NhanVien {
 			stmt.setString(7, nv.getEmail());
 			stmt.setBoolean(8, nv.isGioiTinh());
 			stmt.setBoolean(9, nv.isChucVu());
-			stmt.setDouble(10, nv.getLuong());
+			stmt.setBigDecimal(10, nv.getLuong());
 			n = stmt.executeUpdate();
 		} catch (SQLException e) {
 			e.printStackTrace();
@@ -294,7 +295,7 @@ public class Dao_NhanVien {
 			stmt.setString(6, nv.getEmail());
 			stmt.setBoolean(7, nv.isGioiTinh());
 			stmt.setBoolean(8, nv.isChucVu());
-			stmt.setDouble(9, nv.getLuong());
+			stmt.setBigDecimal(9, nv.getLuong());
 			stmt.setString(10, nv.getMaNhanVien());
 			n = stmt.executeUpdate();
 			dsnv.set(dsnv.indexOf(nv), nv);
@@ -337,7 +338,7 @@ public class Dao_NhanVien {
 				String email = rs.getString(7);
 				boolean gioiTinh = rs.getBoolean(8);
 				boolean chucVu = rs.getBoolean(9);
-				double luong = rs.getDouble(10);
+				BigDecimal luong = rs.getBigDecimal(10);
 				NhanVien nv = new NhanVien(maNV, hoTen, mk, ngaySinh, diaChi, soDienThoai, email, chucVu, gioiTinh,
 						luong);
 				dao.dsnv.add(nv);
@@ -365,7 +366,7 @@ public class Dao_NhanVien {
 				String email = rs.getString(7);
 				boolean gioiTinh = rs.getBoolean(8);
 				boolean chucVu = rs.getBoolean(9);
-				double luong = rs.getDouble(10);
+				BigDecimal luong = rs.getBigDecimal(10);
 				NhanVien nv = new NhanVien(maNV, hoTen, mk, ngaySinh, diaChi, soDienThoai, email, chucVu, gioiTinh,
 						luong);
 				dao.dsnv.add(nv);
@@ -393,7 +394,7 @@ public class Dao_NhanVien {
 				String email = rs.getString(7);
 				boolean gioiTinh = rs.getBoolean(8);
 				boolean chucVu = rs.getBoolean(9);
-				double luong = rs.getDouble(10);
+				BigDecimal luong = rs.getBigDecimal(10);
 				NhanVien nv = new NhanVien(maNV, hoTen, mk, ngaySinh, diaChi, soDienThoai, email, chucVu, gioiTinh,
 						luong);
 				dao.dsnv.add(nv);
@@ -421,7 +422,7 @@ public class Dao_NhanVien {
 				String email = rs.getString(7);
 				boolean gioiTinh = rs.getBoolean(8);
 				boolean chucVu = rs.getBoolean(9);
-				double luong = rs.getDouble(10);
+				BigDecimal luong = rs.getBigDecimal(10);
 				NhanVien nv = new NhanVien(maNV, hoTen, mk, ngaySinh, diaChi, soDienThoai, email, chucVu, gioiTinh,
 						luong);
 				dao.dsnv.add(nv);
